LayoutTests/imported/w3c/web-platform-tests/resize-observer/notify-expected.txt

CONSOLE MESSAGE: line 50: ReferenceError: Can't find variable: ResizeObserver
ResizeObserver tests

t1
t2
t3 inline

Harness Error (FAIL), message = ReferenceError: Can't find variable: ResizeObserver

FAIL ResizeObserver implemented assert_own_property: expected property "ResizeObserver" missing
